requirements.txt
PyQt6==6.5.0
pytesseract==0.3.10
Pillow==10.3.0
python-docx==1.1.0
mss==9.0.1
